MLP DSP: use int16_t instead of int32_t for coeffs
The decoding in read_filter_params insures that coeffs fit within 16bits.

Benchmarks for 2 samples on Arrandale on the whole filter_channel loop:
         A,win32  A,win64    B,win32  B,win64
before:    6110     3172       1870     1068
after:     6100     3160       1862     1026

Aligning the coeffs buffer may also insure that a single line of cache
is accessed, but this had no measurable effect.

+++ libavcodec/mlp.h
@@ -81,7 +81,7 @@ typedef struct FilterParams {
 /** sample data coding information */
 typedef struct ChannelParams {
     FilterParams filter_params[NUM_FILTERS];
-    int32_t     coeff[NUM_FILTERS][MAX_FIR_ORDER];
+    int16_t     coeff[NUM_FILTERS][MAX_FIR_ORDER];
 
     int16_t     huff_offset;      ///< Offset to apply to residual values.
     int32_t     sign_huff_offset; ///< sign/rounding-corrected version of huff_offset
