+++ libavcodec/mlpdsp.c
@@ -24,15 +24,15 @@
 #include "mlpdsp.h"
 #include "mlp.h"
 
-static void mlp_filter_channel(int32_t *state, const int32_t *coeff,
+static void mlp_filter_channel(int32_t *state, const int16_t *coeff,
                                int firorder, int iirorder,
                                unsigned int filter_shift, int32_t mask,
                                int blocksize, int32_t *sample_buffer)
 {
     int32_t *firbuf = state;
     int32_t *iirbuf = state + MAX_BLOCKSIZE + MAX_FIR_ORDER;
-    const int32_t *fircoeff = coeff;
-    const int32_t *iircoeff = coeff + MAX_FIR_ORDER;
+    const int16_t *fircoeff = coeff;
+    const int16_t *iircoeff = coeff + MAX_FIR_ORDER;
     int i;
 
     for (i = 0; i < blocksize; i++) {
